pyproject.toml
[project]
name = "actual-budget-reporter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "actualpy>=0.21.0",
    "google-genai>=1.64.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
